cleaned controllers --clean components + add error handling

+++ server/GroceryInventoryAPI/DTOs/Inventory/PatchInventoryRequest.cs
@@ -23,7 +23,7 @@ public class PatchInventoryRequest
     [Range(0, int.MaxValue, ErrorMessage = "Sales volume must be 0 or greater")]
     public int? SalesVolume { get; set; }
     [Range(0, int.MaxValue, ErrorMessage = "Inventory turnover rate must be 0 or greater")]
-    public int? InventoryTurnoverRate { get; set; }
+    public decimal? InventoryTurnoverRate { get; set; }
     [Range(0, 2, ErrorMessage = "Status must be 0 (Active), 1 (BackOrdered), or 2 (Discontinued)")]
     public Status? Status { get; set; }
     [StringLength(11, MinimumLength = 11, ErrorMessage = "Product ID must be exactly 11 characters")]
